ace_gtv/requirements.txt
# GTV ACE Agent Python依赖 - 基于项目实际需求
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.27.2  # 兼容Python 3.13的版本

# OpenAI/Azure OpenAI集成
openai>=1.0.0

# LangChain/LangGraph依赖 (用于评分Agent)
langchain>=0.1.5
langchain-core>=0.1.5
langchain-openai>=0.1.0
pydantic>=2.0.0

# 文件处理和PDF报告生成
pdfminer.six==20221105
python-docx==1.1.2
reportlab==4.0.4
Pillow==11.0.0  # 兼容Python 3.13的最新版本

# 网页解析
beautifulsoup4>=4.12.0
lxml>=5.0.0

# 数据处理和分析
pandas==2.2.3  # 兼容Python 3.13的版本
numpy>=1.24.0  # 兼容Python 3.13的版本，允许使用2.x版本

# 日志和配置
python-dotenv==1.0.0

# 开发依赖
pytest==7.4.3
pytest-flask==1.3.0

# Web服务和API
werkzeug==3.0.1
